Create a special third-party case for ESRI servers
This currently supports groups (the id in the sources.json *must* be the
group id) and specific FeatureServers.

This should only be used when (a) permission is given for OpenStreetMap
to use the data and (b) the server owner has given their OK. While the
feature calls should be as efficient as possible, many server owners
will be paying for data transfer.

Signed-off-by: Taylor Smock <taylor.smock@kaart.com>