FIXUP: Ensure that a copy of a MapWithAIInfo object actually copies everything.
This caused an issue where an entry would be copied without parameters,
and a failure would occur until the entry list was updated again, at
which point the parameters were set.

This adds a non-regression test that should (hopefully) ensure that this
does not occur again with that specific clone constructor.

Signed-off-by: Taylor Smock <tsmock@fb.com>