ESRISourceReader: Allow for variable max ages
This means that the base ESRI collection urls will be updated more often
(with default settings, every day if JOSM is restarted), while the
specific feature server information urls will be updated weekly (again,
with default settings).

This uses the advanced preference `mirror.maxtime`, which is also used
by CachedFile.

We therefore recommend that anyone thinking about changing
`mirror.maxtime` to for faster updates for the ESRI feature server
collection url to *not* set that value to be less than `14400` (4
hours). This is to avoid hitting public servers too often.

Signed-off-by: Taylor Smock <tsmock@fb.com>