MergeBuildingAddress: Don't delete duplicate addresses when added together
It is possible that the originating source didn't have addr:unit tags,
so other applications such as StreetComplete should detect duplicate
addresses and ask users if there are addr:unit tags, or which one is the
"main" building.

Furthermore, JOSM has a validator for duplicate addresses, so hopefully
the user will see the duplicate addresses and deal with them
appropriately, if it is obvious that one of the buildings is an
outbuilding.

Signed-off-by: Taylor Smock <tsmock@meta.com>